Pass size_t to span functions: components/
In preparation for upgrading various span args from size_t to
StrictNumeric<size_t>, this fixes type mismatches that would cause
compile errors with that change.

Bug: none
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6020501
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Ted Choc <tedchoc@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1382747}

+++ components/qr_code_generator/bitmap_generator.cc
@@ -250,7 +250,7 @@ SkBitmap RenderBitmap(base::span<const uint8_t> data,
 
   // Loop over qr module data and paint to canvas.
   // Paint data modules first, then locators and dino.
-  int data_index = 0;
+  size_t data_index = 0;
   for (int y = 0; y < data_size.height(); y++) {
     for (int x = 0; x < data_size.width(); x++) {
       if (data[data_index++] & 0x1) {
